--- src/main/java/cholog/wiseshop/api/order/dto/ModifyOrderCountRequest.java
@@ -1,5 +1,0 @@
-package cholog.wiseshop.api.order.dto;
-
-public record ModifyOrderCountRequest(Long id,
-                                      int count) {
-}
