+++ src/main/java/cholog/wiseshop/api/order/dto/request/CreateOrderRequest.java
@@ -1,4 +1,4 @@
-package cholog.wiseshop.api.order.dto;
+package cholog.wiseshop.api.order.dto.request;
 
 import cholog.wiseshop.db.order.Order;
 import cholog.wiseshop.db.product.Product;
