+++ src/main/java/cholog/wiseshop/api/order/dto/request/ModifyOrderCountRequest.java
@@ -1,0 +1,4 @@
+package cholog.wiseshop.api.order.dto.request;
+
+public record ModifyOrderCountRequest(int count) {
+}
